Added initialization of proxy contract
