+++ contracts/test/forwarder.js
@@ -1,5 +1,6 @@
 const {reverting} = require('../node_modules/openzeppelin-solidity/test/helpers/shouldFail');
 const createForwarder = require('../helpers/forwarder');
+const { getEncodedCall } = require('../helpers/utils');
 
 const Web3 = require('web3');
 const web3 = new Web3(Web3.givenProvider);
@@ -11,11 +12,11 @@ contract('Forwarder', async (accounts) => {
   it('should be able to use Identity contract with forwarder', async () => {
     const identity = await Identity.new();
     const Forwarder = await createForwarder(identity.address);
-    const forwarder = await Forwarder.new();
+    const encodedData = getEncodedCall(web3, identity, 'initialize', [accounts[0]]);
+    const forwarder = await Forwarder.new(encodedData);
 
     const identityForwarder = await Identity.at(forwarder.address);
-    await identityForwarder.initialise(accounts[0]);
-    await reverting(identityForwarder.initialise(accounts[0]));
+    await reverting(identityForwarder.initialize(accounts[0]));
 
     await identityForwarder.setData('a', 'b');
 
@@ -29,11 +30,11 @@ contract('Forwarder', async (accounts) => {
   it('should be able to use KeyManager contract with forwarder', async () => {
     const keyManager = await KeyManager.new();
     const Forwarder = await createForwarder(keyManager.address);
-    const forwarder = await Forwarder.new();
+    const encodedData = getEncodedCall(web3, keyManager, 'initialize');
+    const forwarder = await Forwarder.new(encodedData);
 
     const keyManagerForwarder = await KeyManager.at(forwarder.address);
-    await keyManagerForwarder.initialise();
-    await reverting(keyManagerForwarder.initialise());
+    await reverting(keyManagerForwarder.initialize());
 
     await keyManagerForwarder.setKey('a', 2, 2);
 
@@ -49,10 +50,8 @@ contract('Forwarder', async (accounts) => {
   it('should not be able to send transaction without function signature', async () => {
     const keyManager = await KeyManager.new();
     const Forwarder = await createForwarder(keyManager.address);
-    const forwarder = await Forwarder.new();
-
-    const keyManagerForwarder = await KeyManager.at(forwarder.address);
-    await keyManagerForwarder.initialise();
+    const encodedData = getEncodedCall(web3, keyManager, 'initialize');
+    const forwarder = await Forwarder.new(encodedData);
 
     // Even if there is not check for function signature in Forwarder contract,
     // this transaction would fail because there is no payable fallback function on KeyManager contract
